requirements.txt
amqp==5.3.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
backcall==0.2.0
billiard==4.2.1
black==23.9.1
celery==5.3.4
certifi==2025.6.15
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.3.2
decorator==5.2.1
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
executing==2.2.0
factory-boy==3.3.0
Faker==37.4.0
flake8==6.1.0
inflection==0.5.1
iniconfig==2.1.0
ipython==8.16.1
isort==5.12.0
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
mccabe==0.7.0

mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==7.4.3
pytest-django==4.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
requests==2.31.0
setuptools<81
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==21.2.0
markdownify==0.11.6
